+++ packages/core/src/components/fields/ArrayField.js
@@ -741,6 +741,8 @@ class ArrayField extends Component {
         return this.renderArrayFieldItem({
           key,
           index,
+          idPrefix,
+          idSeparator,
           canRemove: additional,
           canMoveUp: index >= itemSchemas.length + 1,
           canMoveDown: additional && index < items.length - 1,
@@ -777,6 +779,8 @@ class ArrayField extends Component {
     const {
       key,
       index,
+      idPrefix,
+      idSeparator,
       canRemove = true,
       canMoveUp = true,
       canMoveDown = true,
@@ -819,6 +823,8 @@ class ArrayField extends Component {
           uiSchema={itemUiSchema}
           formData={itemData}
           errorSchema={itemErrorSchema}
+          idPrefix={idPrefix}
+          idSeparator={idSeparator}
           idSchema={itemIdSchema}
           required={this.isItemRequired(itemSchema)}
           onChange={this.onChangeForIndex(index)}